Feature: Refactor attendance processing and add day switch endpoint
Renamed get_assumed_working_time to calc_assumed_working_time and updated all references. Improved working hours adjustment logic and added NIGHT_WORK_START to constants. Enhanced DailyAttendanceRecord with timestamp fields and a custom __init__. Added generate_attendance_record utility and corresponding tests. Refactored day_switch view to use new utilities and handle POST requests, and added the /day_switch URL route.

+++ sao/views.py
@@ -1321,31 +1321,33 @@ def day_switch_time_edit(request, id):
 def day_switch(request):
     """日時切り替え処理"""
 
+    if request.method == "POST":
+        date = request.POST.get("date")
+        date = datetime.datetime.strptime(date, "%Y-%m-%d").date()
+        # WebTimeStampを集めてEmployeeDailyRecordを生成する
+        employees = models.Employee.objects.filter(
+                        user__is_active=True).filter(
+                            join_date__lte=date).filter(
+                            leave_date__gte=date)
+        for employee in employees:
+            # もしEmployeeDailyRecordが存在していたら削除する
+            models.EmployeeDailyRecord.objects.filter( 
+                employee=employee, date=date).delete()
+            # WebTimeStampを集める
+            stamps = utils.collect_webstamp(employee, date)
+            # EmployeeDailyRecordを生成する
+            utils.generate_daily_record(stamps, employee, date)
 
+            # EmployeeDailyRecordを集めてDailyAttendanceRecordを生成する
+            records = [ x for x in models.EmployeeDailyRecord.objects.filter(employee=employee).filter(
+                date=date) ]
+            if len(records) > 1:
+                raise Exception(f"同日に複数の勤務記録が存在しています: {employee} {date}")
+            
+            # DailyAttendanceRecordを生成する
+            utils.generate_attendance_record(records[0])
 
-
-
-    date = datetime.date.today()
-    # WebTimeStampを集めてEmployeeDailyRecordを生成する
-    employees = models.Employee.objects.filter(
-                    user__is_active=True).filter(
-                        join_date__lte=date).filter(
-                        leave_date__gte=date)
-    for employee in employees:
-        # もしEmployeeDailyRecordが存在していたら削除する
-        models.EmployeeDailyRecord.objects.filter( 
-            employee=employee, date=date).delete()
-        # WebTimeStampを集める
-        stamps = [x.stamp for x in models.WebTimeStamp.objects.filter(employee=employee).filter(
-            date=date).order_by("stamp") if x.stamp is not None]
-        # EmployeeDailyRecordを生成する
-        utils.generate_daily_record(stamps, employee, date)
-
-        # EmployeeDailyRecordを集めてDailyAttendanceRecordを生成する
-        records = [ x for x in models.EmployeeDailyRecord.objects.filter(employee=employee).filter(
-            date=date) ]
-        # utils.generate_attendance_record(records, employee, date)
-
+    return HttpResponse("day switch done")
